DEX-671 move flattenTree to utility file and fix bug therein where duplicate leaves were being added during tree flattening
