+++ src/hooks/useOptions.js
@@ -1,21 +1,8 @@
 import { useMemo } from 'react';
-import { cloneDeep, get } from 'lodash-es';
+import { get } from 'lodash-es';
 
+import { flattenTree } from '../utils/formatters';
 import useSiteSettings from '../models/site/useSiteSettings';
-
-function flattenTree(regions) {
-  const flatTree = cloneDeep(regions);
-
-  function addLevel(leaves) {
-    leaves.forEach(leaf => {
-      flatTree.push(leaf);
-      if (leaf.locationID) addLevel(leaf.locationID);
-    });
-  }
-
-  addLevel(regions);
-  return flatTree;
-}
 
 export default function useOptions() {
   const {
